playwright specs on mock activity feed
